+++ addOns/alertFilters/CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 ### Changed
+- Allow to filter by alert reference (Issue 7438).
 - Maintenance changes.
 
 ### Fixed